// typescript/user-interface/packages/ui-app/src/ts/components/analyst-ui/components/map/types.ts
export interface MapEventSource extends EventRow {
  readonly id: string;
  readonly edgeEventType: EdgeTypes;
  readonly isOpen: boolean;
  readonly conflict: boolean;
  readonly time: number;
  readonly latitudeDegrees: number;
  readonly longitudeDegrees: number;
  readonly depthKm: number;
  readonly region: string;
  readonly confidenceSemiMajorAxis: number;
  readonly confidenceSemiMinorAxis: number;
  readonly coverageSemiMajorAxis: number;
  readonly coverageSemiMinorAxis: number;
  readonly coverageSemiMajorAxisTrend?: number;
  readonly confidenceSemiMajorAxisTrend?: number;
  readonly magnitudeMb: number;
  readonly magnitudeMs: number;
  readonly magnitudeMl: number;
  readonly activeAnalysts: string[];
  readonly preferred: string;
  readonly status: EventTypes.EventStatus;
}

/**
 * Values to be added to signal detection entity properties
 */
export interface MapSDEntityValues {
  readonly detectionTime: {
    detectionTimeValue: string;
    detectionTimeUncertainty: number;
  };
  readonly azimuth: {
    azimuthValue: number;
    azimuthUncertainty: number;
  };
  readonly slowness: {
    slownessValue: number;
    slownessUncertainty: number;
  };
  readonly phaseValue: SignalDetectionTypes.PhaseTypeMeasurementValue;
  readonly associatedEventTimeValue: string;
  readonly signalDetectionColor: string;
  readonly status: string;
  readonly edgeSDType: EdgeTypes;
  readonly stationName: string;
}